backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djoser==2.2.0
psycopg2-binary==2.9.9
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
django-filter==23.4
drf-extra-fields==3.7.0
django-colorfield==0.10.1
tqdm==4.67.1